refactor(remote): split the group domain out of the daemon monolith
daemon.rs was a 1676-line monolith mixing the run loop, wire routing,
peer networking, and the entire group-mode domain. Split it into three
focused modules so each concern is cohesive and the boundaries are
explicit:

- context.rs: DaemonContext shared state
- net.rs: iroh dial/exchange primitives (exchange, exchange_with_peer,
  connect_default, observed_endpoint), taking (store, endpoint) instead
  of the whole context
- group.rs: the group domain — roster reconciliation, sync, broadcast,
  join/leave/kick, query fan-out — with its tests; the daemon keeps
  only the run loop and thin wire routing

Behavioral deltas (all intentional, none compat shims):
- GroupMemberAdded now converges through the shared reconcile_roster
  (idempotent upsert + stale-share prune + grant) instead of a bespoke
  add-only loop
- GroupSync serves roster_for, which encodes the membership gate
  (unknown group / non-member / full roster) once
- the dead empty-targets special case in GroupQuery is gone: an
  existing group always has its owner as a member and group_fan_out
  already returns an empty result on an empty corpus
- GroupJoin drops the expect() by folding the membership check into
  Option::filter

No re-exports, no wrappers: moved code was deleted from daemon.rs.
